refactor: extract shared SyncStatusPopover to @epicenter/svelte
Deduplicates the near-identical SyncStatusIndicator.svelte across fuji,
opensidian, and tab-manager into a single shared component. Honeycrisp's
AccountPopover (which lacked sync status awareness) is replaced too—it
now gets the full sync UI for free.

The shared component accepts auth, workspace, syncNoun, and
onSocialSignIn as props. Each app passes its own auth form config.

Also documents the onLogout contract in create-auth.svelte.ts—the hook
is not awaited (subscription callback limitation), and the type signature
now reflects this honestly as () => void | Promise<void>.
