+++ apps/honeycrisp/src/lib/components/Sidebar.svelte
@@ -1,10 +1,11 @@
 <script lang="ts">
+	import { SyncStatusPopover } from '@epicenter/svelte/sync-status-popover';
 	import * as Collapsible from '@epicenter/ui/collapsible';
 	import * as Sidebar from '@epicenter/ui/sidebar';
 	import FileTextIcon from '@lucide/svelte/icons/file-text';
 	import PlusIcon from '@lucide/svelte/icons/plus';
 	import TrashIcon from '@lucide/svelte/icons/trash-2';
-	import AccountPopover from '$lib/components/AccountPopover.svelte';
+	import { auth, workspace } from '$lib/client';
 	import FolderMenuItem from '$lib/components/FolderMenuItem.svelte';
 	import { foldersState, notesState, viewState } from '$lib/state';
 </script>
@@ -14,7 +15,16 @@
 		<div class="flex items-center justify-between px-2 py-1">
 			<span class="text-sm font-semibold">Honeycrisp</span>
 			<div class="flex items-center gap-1">
-				<AccountPopover />
+				<SyncStatusPopover
+					{auth}
+					{workspace}
+					syncNoun="notes"
+					onSocialSignIn={() =>
+						auth.signInWithSocialRedirect({
+							provider: 'google',
+							callbackURL: window.location.origin,
+						})}
+				/>
 				<Sidebar.Trigger />
 			</div>
 		</div>
