+++ packages/svelte-utils/package.json
@@ -17,7 +17,8 @@
 		".": "./src/index.ts",
 		"./auth": "./src/auth/index.ts",
 		"./auth-form": "./src/auth-form/index.ts",
-		"./workspace-gate": "./src/workspace-gate/index.ts"
+		"./workspace-gate": "./src/workspace-gate/index.ts",
+		"./sync-status-popover": "./src/sync-status-popover/index.ts"
 	},
 	"dependencies": {
 		"@epicenter/constants": "workspace:*",
